Code for CS3244 Project
tgs_lr, tgs_svm and tgs_cnn_class were all written by me. These 3 are used as baseline models.

tgs_kernel is a U-net implementation following the public kernel by Peter Hönigschmid. The purpose of the code was more of a model for learning the U-net model and understanding how the hyper parameters work in affecting the outcome of a NN.